+++ src/functions/filter-array.ts
@@ -2,4 +2,4 @@
  * Filter 'undefined' elements from an array.
  * @param array - The array to filter.
  */
-export const filterArray = <T>(array: Array<T | undefined>): T[] => array.filter((e): e is T => e !== undefined);
+export const filterArray = <T>(array: Array<T | undefined>): T[] => array.filter((element): element is T => element !== undefined);