+++ src/functions/sort-array-by-object-property-length.ts
@@ -5,7 +5,7 @@
  * @param inverse - Inverse the output (descending).
  */
 export const sortArrayByObjectPropertyLength = <T>(
-    array: T[], attributeGetter: (item: T) => { length: number; }, inverse = false
+    array: T[], attributeGetter: (item: T) => { length: number; }, inverse = false,
 ): void => {
 
     const compare = (a: any, b: any) => {