+++ src/functions/split-array-into-chunks.ts
@@ -10,8 +10,8 @@ export const splitArrayIntoChunks = <T>(array: T[], chunkSize: number): T[][] =>
     }
 
     const chunks: T[][] = [];
-    for (let i = 0; i < array.length; i += chunkSize) {
-        chunks.push(array.slice(i, i + chunkSize));
+    for (let index = 0; index < array.length; index += chunkSize) {
+        chunks.push(array.slice(index, index + chunkSize));
     }
     return chunks;
 };